chore: refactor component into new file
